Adds new daily_mailer action with small refactor.
