+++ mailers/mailer.py
@@ -3,16 +3,18 @@ from email.mime.multipart import MIMEMultipart
 from email.mime.text import MIMEText
 
 from config.secrets import GMAIL
+from mailers.helpers import format_bookmark
 
 
 SMTP_HOST = 'smtp.gmail.com'
 SMTP_PORT = 587
+DEFAULT_SUBJECT = 'Daily Pinboard Bulletin'
 
 
 class Mailer:
-    def __init__(self, bookmarks):
+    def __init__(self, bookmarks, **keywords):
         self.bookmarks = bookmarks
-        self.subject = 'Daily Pinboard Bulletin'
+        self.subject = keywords.get('subject', DEFAULT_SUBJECT)
         self.from_name = 'Pinprick Bot'
         self.smtp_address = GMAIL['address']
         self.smtp_pass = GMAIL['password']
@@ -50,37 +52,6 @@ class Mailer:
 
 {}
 """
-        bookmark_blocks = [self.format_bookmark(bookmark) for bookmark in bookmarks]
+        bookmark_blocks = [format_bookmark(bookmark) for bookmark in bookmarks]
         bookmark_block = "\n".join(bookmark_blocks)
         return email_f.format(len(bookmarks), bookmark_block)
-
-    def format_bookmark(self, bookmark):
-        bookmark_f = """\
-<div class="bookmark" style="margin-bottom:4px">
-  <h4 style="margin-bottom:4px">
-    <a href="{}">{}</a>
-  </h4>
-  <div class="tags">
-    {}
-  </div>
-  <div class="meta">
-    {}
-  </div>
-</div>"""
-
-        def format_tags(bookmark):
-            tag_blocks = [format_tag(bookmark, tag) for tag in bookmark.tags]
-            return ', '.join(tag_blocks)
-
-        def format_tag(bookmark, tag):
-            tag_f = '<a href="{}">{} ({})</a>'
-            return tag_f.format(bookmark.tag_to_url(tag), tag.name, tag.count)
-
-        def format_meta(bookmark):
-            meta_f = '<span>{}</span> &bull; <a href="{}">{}</a>'
-            return meta_f.format(bookmark.created_at, bookmark.service_url, bookmark.service_url)
-
-        return bookmark_f.format(bookmark.url,
-                                 bookmark.title,
-                                 format_tags(bookmark),
-                                 format_meta(bookmark))
